feat: gqlgen plugin to add schema from engagement library
