+++ go.mod
@@ -21,7 +21,7 @@ require (
 	github.com/kevinburke/go-types v0.0.0-20210723172823-2deba1f80ba7 // indirect
 	github.com/markbates/pkger v0.17.1
 	github.com/savannahghi/converterandformatter v0.0.11
-	github.com/savannahghi/engagement v0.0.4
+	github.com/savannahghi/engagement v0.0.5
 	github.com/savannahghi/enumutils v0.0.3
 	github.com/savannahghi/errorcodeutil v0.0.3
 	github.com/savannahghi/feedlib v0.0.4
@@ -33,6 +33,7 @@ require (
 	github.com/segmentio/ksuid v1.0.4
 	github.com/sirupsen/logrus v1.8.1
 	github.com/stretchr/testify v1.7.0
+	github.com/vektah/gqlparser v1.3.1
 	github.com/vektah/gqlparser/v2 v2.1.0
 	gitlab.slade360emr.com/go/commontools/crm v0.0.0-20210820122023-7a581f80cb5f // indirect
 	go.opencensus.io v0.23.0
