--- pkg/engagement/presentation/graph/fcm.graphql
@@ -1,28 +1,0 @@
-extend type Mutation {
-    sendNotification(
-        registrationTokens: [String!]!,
-        data: Map!,
-        notification: FirebaseSimpleNotificationInput!,
-        android: FirebaseAndroidConfigInput,
-        ios: FirebaseAPNSConfigInput,
-        web: FirebaseWebpushConfigInput
-    ): Boolean!
-
-    sendFCMByPhoneOrEmail(
-        phoneNumber: String,
-        email: String,
-        data: Map!
-        notification: FirebaseSimpleNotificationInput!,
-        android: FirebaseAndroidConfigInput,
-        ios: FirebaseAPNSConfigInput,
-        web: FirebaseWebpushConfigInput
-    ):Boolean!
-}
-
-extend type Query {
-    notifications(
-        registrationToken: String!
-        newerThan: Time!
-        limit: Int!
-    ): [SavedNotification!]!
-}
