--- pkg/engagement/presentation/graph/inputs.graphql
@@ -1,23 +1,0 @@
-
-input FirebaseSimpleNotificationInput {
-    title: String!
-    body: String!
-    imageURL: String!
-    data: Map
-}
-
-input FirebaseAndroidConfigInput {
-    priority: String!
-    collapseKey: String
-    restrictedPackageName: String
-    data: Map
-}
-
-input FirebaseWebpushConfigInput {
-    headers: Map
-    data: Map
-}
-
-input FirebaseAPNSConfigInput {
-    headers: Map
-}
