--- pkg/engagement/presentation/graph/sms.graphql
@@ -1,25 +1,0 @@
-extend type Mutation {
-  send(to: String!, message: String!): SendMessageResponse!
-
-  sendToMany(message: String!, to: [String!]!): SendMessageResponse!
-}
-
-type Recipient {
-  number: String!
-  cost: String!
-  status: String!
-  messageID: String!
-}
-
-type SMS {
-  recipients: [Recipient!]!
-}
-
-type SendMessageResponse {
-  SMSMessageData: SMS!
-}
-
-enum SenderID {
-  SLADE360
-  BEWELL
-}
