--- pkg/engagement/presentation/graph/types.graphql
@@ -1,49 +1,0 @@
-"""
-AccessToken is used to return the credentials that are needed in order
-to access a Twilio video room.
-"""
-type AccessToken @key(fields: "jwt") @key(fields: "uniqueName") {
-  jwt: String!
-  uniqueName: String!
-  sid: String!
-  dateUpdated: Time!
-  status: String!
-  type: String!
-  maxParticipants: Int!
-  duration: Int
-}
-
-type FirebaseSimpleNotification {
-  title: String!
-  body: String!
-  imageURL: String
-  data: Map
-}
-
-type FirebaseAndroidConfig {
-  collapseKey: String!
-  priority: String!
-  restrictedPackageName: String!
-  data: Map
-}
-
-type FirebaseWebpushConfig {
-  headers: Map
-  data: Map
-}
-
-type FirebaseAPNSConfig {
-  headers: Map
-}
-
-type SavedNotification @key(fields: "id") {
-  id: String!
-  registrationToken: String!
-  messageID: String!
-  timestamp: Time!
-  data: Map
-  notification: FirebaseSimpleNotification
-  androidConfig: FirebaseAndroidConfig
-  webpushConfig: FirebaseWebpushConfig
-  apnsConfig: FirebaseAPNSConfig
-}
