+++ pkg/engagement/usecases/usecases.go
@@ -4,12 +4,12 @@ import (
 	"context"
 
 	"github.com/savannahghi/engagement/pkg/engagement/infrastructure"
-	openSource "github.com/savannahghi/engagement/pkg/engagement/usecases"
+	engagementLib "github.com/savannahghi/engagement/pkg/engagement/usecases"
 )
 
 // Usecases is an interface that combines of all usescases
 type Usecases interface {
-	openSource.Usecases
+	engagementLib.Usecases
 	// New usecases
 	// Just an example of how you can add new usecase
 	TestFeature(ctx context.Context) (bool, error)
@@ -18,14 +18,14 @@ type Usecases interface {
 //CombinedInteractor is used to combine both our internal and open source
 // usecases implementation
 type CombinedInteractor struct {
-	openSource.Usecases
+	engagementLib.Usecases
 	InternalInteractor
 }
 
 // NewUsecasesInteractor initializes new combined usecases interactor
 func NewUsecasesInteractor(infrastructure infrastructure.Infrastructure) *CombinedInteractor {
 
-	open := openSource.NewUsecasesInteractor(infrastructure)
+	open := engagementLib.NewUsecasesInteractor(infrastructure)
 	internal := NewInternalInteractor(infrastructure)
 
 	impl := &CombinedInteractor{
@@ -38,12 +38,18 @@ func NewUsecasesInteractor(infrastructure infrastructure.Infrastructure) *Combin
 
 // InternalInteractor ...
 type InternalInteractor struct {
+	engagementLib  engagementLib.Usecases
 	infrastructure infrastructure.Infrastructure
 }
 
 // NewInternalInteractor ...
 func NewInternalInteractor(infrastructure infrastructure.Infrastructure) *InternalInteractor {
-	return &InternalInteractor{infrastructure: infrastructure}
+	library := engagementLib.NewUsecasesInteractor(infrastructure)
+
+	return &InternalInteractor{
+		infrastructure: infrastructure,
+		engagementLib:  library,
+	}
 }
 
 // TestFeature ...
